% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_cameras.R
\name{align_cameras}
\alias{align_cameras}
\title{Align camera axis}
\usage{
align_cameras(flref, fvref, output, center = c(0, 0), zoom = 1,
  autopos = T)
}
\arguments{
\item{obj}{A target image of Image object or an array.}

\item{ref}{A reference image of Image object or an array.}
}
\description{
Align camera axis
}
\examples{
align_cameras()

}

