% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_trajectories.R
\name{analyze_trajectories}
\alias{analyze_trajectories}
\title{Analyze trajectory of flies}
\usage{
analyze_trajectories(dir, output, fpsfv, interaction = F)
}
\arguments{
\item{dir}{Input directory.}

\item{output}{Output path for saving results.}

\item{fpsfv}{fly-view frame fate.}

\item{interaction}{logical. If True, measures distance between two flies.}
}
\description{
Analyze trajectory of flies
}
\examples{
analyze_trajectories()

}
