% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_angle.R
\name{detect_angle}
\alias{detect_angle}
\title{Detect angle of the beads}
\usage{
detect_angle(img)
}
\arguments{
\item{img}{A binary labelled image or array.}
}
\description{
Detect angle of the beads
}
\examples{
detect_angle()

}
