% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_interaction.R
\name{detect_interaction}
\alias{detect_interaction}
\title{Detect fly-fly interactions}
\usage{
detect_interaction()
}
\arguments{
\item{obj}{A target image of Image object or an array.}

\item{ref}{A reference image of Image object or an array.}
}
\description{
Detect fly-fly interactions
}
\examples{
detect_interaction()

}

