% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_window.R
\name{detect_window}
\alias{detect_window}
\title{Perform image processing to segment the imaging window}
\usage{
detect_window()
}
\arguments{
\item{obj}{A target image of Image object or an array.}

\item{ref}{A reference image of Image object or an array.}
}
\description{
Perform image processing to segment the imaging window
}
\examples{
detect_window()
}

