% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dF_F0_image.R
\name{dF_F0_image}
\alias{dF_F0_image}
\title{Create delta F over F pseudocolor representation}
\usage{
dF_F0_image(flimgreg, fvimgbwbrfhregimg, regimgi, colmax = F, cmin, cmax,
  goodfr = NULL, output)
}
\arguments{
\item{obj}{A target image of Image object or an array.}

\item{ref}{A reference image of Image object or an array.}
}
\description{
Create delta F over F pseudocolor representation
}
\examples{
dF_F0_image()

}

