% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_goodframes.R
\name{find_goodframes}
\alias{find_goodframes}
\title{Calculate window size, position, motion, and blurriness}
\usage{
find_goodframes(window_mask, fvimgl, output, motion_thresh = 10,
  dist_thresh = 20, sharp_thresh = 150, reuse = F)
}
\arguments{
\item{obj}{A target image of Image object or an array.}

\item{ref}{A reference image of Image object or an array.}
}
\description{
Calculate window size, position, motion, and blurriness
}
\examples{
find_goodframes()

}

