% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flyception.R
\name{Flyception}
\alias{Flyception}
\title{FlyceptionR main script}
\usage{
Flyception(rdir, dir, prefix, autopos = T, video_out = F, interaction = T,
  stimulus = 100, stimlen = 200, stimplotlen = 800, reuse = T,
  fmf2tif = T, fpsfl = 100, zoom = 0.85, FOI = F, binning = 1)
}
\arguments{
\item{obj}{A target image of Image object or an array.}

\item{ref}{A reference image of Image object or an array.}
}
\description{
FlyceptionR main script
}
\examples{
Flyception()

}

