% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageJ_crop_append.R
\name{imageJ_crop_append}
\alias{imageJ_crop_append}
\title{Crop and concatenate fluo_view image with up to 2 channels using ImageJ}
\usage{
imageJ_crop_append(dir, ch = 1, roi = c(383, 0, 256, 256))
}
\arguments{
\item{dir}{path to the directory that contains the data}

\item{ch}{channel name used for output files}

\item{roi}{roi coordinates (x, y, w, h), note origin is (0,0)}
}
\description{
Crop and concatenate fluo_view image with up to 2 channels using ImageJ
}
\examples{
imageJ_crop_append()
}
