% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_ROI.R
\name{measureROI}
\alias{measureROI}
\title{Create delta F over F plot for ROI}
\usage{
measureROI(img, mask, output)
}
\arguments{
\item{obj}{A target image of Image object or an array.}

\item{ref}{A reference image of Image object or an array.}
}
\description{
Create delta F over F plot for ROI
}
\examples{
measureROI()

}

