% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dF_F0.R
\name{plot_dF_F0}
\alias{plot_dF_F0}
\title{Plot delta F over F}
\usage{
plot_dF_F0(intensity, goodfr)
}
\arguments{
\item{obj}{A target image of Image object or an array.}

\item{ref}{A reference image of Image object or an array.}
}
\description{
Plot delta F over F
}
\examples{
plot_dF_F0()

}
