% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_images.R
\name{register_images}
\alias{register_images}
\title{Perform image registration for fly-view, window, and fluo-view}
\usage{
register_images(fvimgl, flimgrt, fvimgbwbrfh, angles, output, zoom, center,
  reuse = F)
}
\arguments{
\item{obj}{A target image of Image object or an array.}

\item{ref}{A reference image of Image object or an array.}
}
\description{
Perform image registration for fly-view, window, and fluo-view
}
\examples{
register_images()

}

