% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_trajectories.R
\name{analyze_trajectories}
\alias{analyze_trajectories}
\title{Analyze trajectory of flies}
\usage{
analyze_trajectories(dir, output, fpsfv, interaction = F)
}
\arguments{
\item{obj}{A target image of Image object or an array.}

\item{ref}{A reference image of Image object or an array.}
}
\description{
Analyze trajectory of flies
}
\examples{
analyze_trajectories()

}
