% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_flash.R
\name{detect_flash}
\alias{detect_flash}
\title{Detect flash}
\usage{
detect_flash(input, output, type = c("fluo", "fly", "arena"), flash_thresh,
  reuse = F)
}
\arguments{
\item{obj}{A target image of Image object or an array.}

\item{ref}{A reference image of Image object or an array.}
}
\description{
Detect flash
}
\examples{
detect_flash()

}
