% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_frames.R
\name{sync_frames}
\alias{sync_frames}
\title{Synchronize camera frames}
\usage{
sync_frames(dir, fluo_flash, fly_flash, arena_flash, output, reuse = F)
}
\arguments{
\item{obj}{A target image of Image object or an array.}

\item{ref}{A reference image of Image object or an array.}
}
\description{
Synchronize camera frames
}
\examples{
sync_frames()
}
